Add an optimized topology file reading routine (C++) that makes reading topology files in ParmEd between 3 and 5 times faster. The bindings are written using the pure Python C API (rather than Cython or Boost.Python or SWIG), and is entirely optional (if the Python headers are not available, the extension is not built and ParmEd falls back to the original pure Python implementation).
Note that because ParmEd makes _no_ assumptions about which flags are available,
or how much data is in each section, the parsing can't be made as fast as, for
instance, cpptraj.  In my test of the 2 million atom system, the pure Python
parser took 58 seconds to parse the file, this new parser took 13 seconds, and
cpptraj took 7.5 seconds.

More common topology files took ~10 to 15 seconds to parse; with this optimized
parser they will take more like 2 to 5 seconds.  prmtop _writing_ is much
simpler (you always know how much data you're writing), and usually takes a
fraction of the time that reading takes -- there is no advantage to trying to
optimize that process (writing with pure Python is already faster than reading
with C++, I believe).